requirements.txt
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
Flask==3.1.2
flask-restful==0.3.10
marshmallow==4.2.0